refactor(client): common error type
